Separate our sensormath functions from SensorUtils.cpp.
